OrderQueue and cancelled order changes

+++ src/main/java/com/wityorestaurant/modules/reservation/repository/ReservationRepository.java
@@ -19,5 +19,5 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
     public Reservation getByCustomerId(@Param("customerInfo") String customerInfo,@Param("restId") Long restId);
 
     @Query(value = "SELECT * from reservation r inner join resttable c on c.id=r.table_Id where c.table_Number=:tableId and c.rest_id=:restId", nativeQuery = true)
-    public List<Reservation> getByTableId(@Param("tableId") Integer tableId,@Param("restId") Long restId);
+    public List<Reservation> getByTableId(@Param("tableId") Long tableId,@Param("restId") Long restId);
 }
